+++ src/io/graphviz.i.hpp
@@ -207,7 +207,7 @@ void deserialize(std::istream &in, G &graph,
           inserted_vertices[source_vertex_name] = source;
         }
 
-        for (size_t i = 1; i < edge_vertices.size(); ++i) {
+        for (int64_t i = 1; i < edge_vertices.size(); ++i) {
           std::string target_vertex_name = utils::trim(edge_vertices[i]);
           if (!target_vertex_name.empty()) {
